+++ src/core/object/watch/CHANGELOG.md
@@ -14,6 +14,7 @@ Changelog
 #### :bug: Bug Fix
 
 * Fixed setting new properties of a watchable object by a deep path using the accessor-based engine
+* Now watchable Map uses the `entries` iterator by default
 
 ## v3.79.0 (2022-03-18)