+++ src/core/object/watch/CHANGELOG.md
@@ -9,7 +9,7 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v3.??.?? (2022-??-??)
+## v3.87.1 (2022-08-15)
 
 #### :bug: Bug Fix